requirements.txt
numpy==1.26.4
pandas==2.2.2
pyarrow==16.1.0
fastparquet==2024.5.0
scikit-learn==1.5.1
shap==0.45.1
typer==0.12.3
click==8.1.7  # fix Typer help crash
joblib==1.4.2
PyYAML==6.0.2
matplotlib==3.9.0
streamlit==1.37.1

google-generativeai>=0.7.0
python-dotenv>=1.0.1
requests>=2.32.0